+++ packages/flutter/lib/src/locales/de.dart
@@ -45,6 +45,7 @@ const PlassLabels de = PlassLabels(
   maximize: 'Maximieren',
   restore: 'Wiederherstellen',
   resizeWindow: 'Fenstergröße ändern',
+  moveWindow: 'Fenster verschieben',
   overlay: 'Overlay',
   pagination: 'Seitennavigation',
   paginationPrevious: 'Vorherige Seite',