+++ packages/flutter/lib/src/locales/ko.dart
@@ -45,6 +45,7 @@ const PlassLabels ko = PlassLabels(
   maximize: '최대화',
   restore: '이전 크기로',
   resizeWindow: '창 크기 조절',
+  moveWindow: '창 이동',
   overlay: '오버레이',
   pagination: '페이지 이동',
   paginationPrevious: '이전 페이지',