+++ packages/react/src/locales/ja.ts
@@ -45,6 +45,7 @@ export const ja: PlassLabels = {
   maximize: '最大化',
   restore: '元に戻す',
   resizeWindow: 'ウィンドウのサイズを変更',
+  moveWindow: 'ウィンドウを移動',
   overlay: 'オーバーレイ',
   notifications: '通知',
   pagination: 'ページ送り',