+++ packages/react/src/locales/zh-Hans.ts
@@ -45,6 +45,7 @@ export const zhHans: PlassLabels = {
   maximize: '最大化',
   restore: '还原',
   resizeWindow: '调整窗口大小',
+  moveWindow: '移动窗口',
   overlay: '遮罩',
   notifications: '通知',
   pagination: '分页',